The server configurations with Node js and Tyopescripit.
The server is initialized and the Socket.io is set to receive and emit the events for bidirectional communication with the client. The config file contains the server configurations for the port. Also the tests for the socket.io inside the src folder.

The README file contains more information about running the server and tests