The client side
Built with React js and Typescript as the template. It contains several folders and files. Created a custom hook for the socket in the client side, and component for Editor.
More information on how to run the client check on the README file under the folder.
The Editor creates the client socket and we are able to emit and receive socket events to and from the sever on port 3000.